Added table name does not exist exception

+++ SQLModel/SQLConnection.py
@@ -14,13 +14,28 @@ class SQLConnection:
             self.sql_execute("./SQLScript/sql_init.sql")
         self.cursor.execute(f"USE {self.schema_name}")
 
+    def sql_get_table(self):
+        self.cursor.execute("SELECT DISTINCT TABLE_NAME FROM"
+                            " information_schema.TABLE_CONSTRAINTS"
+                            " WHERE CONSTRAINT_SCHEMA = 'Local_Library_Schema'")
+        tables = []
+        for table in self.cursor.fetchall():
+            tables.append(table[0])
+        return tables
+
     def sql_get_pk(self, table_name):
-        self.cursor.execute("SELECT COLUMN_NAME "
-                            "FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE "
-                            f"WHERE TABLE_SCHEMA = '{self.schema_name}' "
-                            "AND CONSTRAINT_NAME = 'PRIMARY' "
-                            f"AND TABLE_NAME = '{table_name}';")
-        return self.cursor.fetchone()[0]
+        try:
+            if table_name not in self.sql_get_table():
+                raise Exception("table name does not match any existing tables")
+            self.cursor.execute("SELECT COLUMN_NAME "
+                                "FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE "
+                                f"WHERE TABLE_SCHEMA = '{self.schema_name}' "
+                                "AND CONSTRAINT_NAME = 'PRIMARY' "
+                                f"AND TABLE_NAME = '{table_name}';")
+            return self.cursor.fetchone()[0]
+        except Exception as msg:
+            print("ERROR", msg)
+            return False
 
     def sql_execute(self, filename):
         fd = open(filename, 'r')
@@ -37,6 +52,8 @@ class SQLConnection:
 
     def sql_select(self, table_name=str, fields=None):
         try:
+            if table_name not in self.sql_get_table():
+                raise Exception("table name does not match any existing tables")
             if fields:
                 if isinstance(fields, tuple) or isinstance(fields, set) or isinstance(fields, list):
                     fields = ", ".join(fields)
@@ -44,20 +61,28 @@ class SQLConnection:
             else:
                 self.cursor.execute(f"SELECT * FROM {table_name}")
             return self.cursor.fetchall()
-        except (mysql.connector.errors.ProgrammingError, TypeError) as msg:
+        except (mysql.connector.errors.ProgrammingError, TypeError, Exception) as msg:
             print("ERROR", msg)
             return False
 
     def sql_select_joint_fk(self, table_name=str):
         try:
+            if table_name not in self.sql_get_table():
+                raise Exception("table name does not match any existing tables")
+            if table_name == 'BOOKS':
+                pass
+            elif table_name == '':
+                pass
             self.cursor.execute(f"SELECT * FROM {table_name}")
             return self.cursor.fetchall()
-        except (mysql.connector.errors.ProgrammingError, TypeError) as msg:
+        except (mysql.connector.errors.ProgrammingError, TypeError, Exception) as msg:
             print("ERROR", msg)
             return False
 
     def sql_insert(self, table_name=str, data_dict=dict):
         try:
+            if table_name not in self.sql_get_table():
+                raise Exception("table name does not match any existing tables")
             key_list = []
             value_list = []
             for col in data_dict:
@@ -70,28 +95,34 @@ class SQLConnection:
                 self.cursor.execute(f"INSERT INTO {table_name} ({key_list}) VALUES {tuple(value_list)}")
             self.connection.commit()
             return True
-        except (mysql.connector.errors.ProgrammingError, mysql.connector.errors.IntegrityError, TypeError) as msg:
+        except (mysql.connector.errors.ProgrammingError, mysql.connector.errors.IntegrityError,
+                Exception, TypeError) as msg:
             print("ERROR", msg)
             return False
 
     def sql_update(self, table_name=str, primary_key=int, data_dict=dict):
         try:
+            if table_name not in self.sql_get_table():
+                raise Exception("table name does not match any existing tables")
             pk_col = self.sql_get_pk(table_name)
             for col in data_dict:
                 self.cursor.execute(f"UPDATE {table_name} SET {col} = '{data_dict[col]}' WHERE {pk_col} = '{primary_key}'")
             self.connection.commit()
             return True
-        except (mysql.connector.errors.ProgrammingError, mysql.connector.errors.IntegrityError, TypeError) as msg:
+        except (mysql.connector.errors.ProgrammingError, mysql.connector.errors.IntegrityError,
+                Exception, TypeError) as msg:
             print("ERROR", msg)
             return False
 
     def sql_delete(self, table_name=str, primary_key=int):
         try:
+            if table_name not in self.sql_get_table():
+                raise Exception("table name does not match any existing tables")
             pk_col = self.sql_get_pk(table_name)
             self.cursor.execute(f"DELETE FROM {table_name} WHERE {pk_col} = {primary_key}")
             self.connection.commit()
             return True
-        except (mysql.connector.errors.ProgrammingError, mysql.connector.errors.IntegrityError) as msg:
+        except (mysql.connector.errors.ProgrammingError, mysql.connector.errors.IntegrityError, Exception) as msg:
             print("ERROR", msg)
             return False
